feat(1sqlite): median/percentile + carray (флаги 3.51, VC6-патч)
Включены SQLITE_ENABLE_PERCENTILE (median/percentile/percentile_cont/disc)
и SQLITE_ENABLE_CARRAY — встроены в амальгаму с 3.51, авто-регистрация.
VC6-патч в percentile.c: ix = rPct*(double)(i64)(nUsed-1) — обход C2520
(unsigned __int64 -> double). carray скомпилирован, но из чистого SQL 1С
недоступен (нужен sqlite3_bind_pointer из C). README дополнен.

+++ SQL_DBF/SQLite/sqlite3.c
@@ -136915,7 +136915,7 @@ static void percentCompute(sqlite3_context *pCtx, int bIsFinal){
       percentSort(p->a, p->nUsed);
       p->bSorted = 1;
     }
-    ix = p->rPct*(p->nUsed-1);
+    ix = p->rPct*(double)(i64)(p->nUsed-1);  /* VC6: u64->double via i64 (percentile) */
     i1 = (unsigned)ix;
     if( settings & 1 ){
       vx = p->a[i1];
